Fix cStringIO ssize_t bugs.
git-svn-id: http://svn.python.org/projects/python/branches/ssize_t@41871 6015fed2-1504-0410-9fe1-9d1591cc4771

+++ Include/cStringIO.h
@@ -29,7 +29,7 @@ static struct PycStringIO_CAPI {
  /* Read a string from an input object.  If the last argument
     is -1, the remainder will be read.
     */
-  int(*cread)(PyObject *, char **, int);
+  int(*cread)(PyObject *, char **, Py_ssize_t);
 
  /* Read a line from an input object.  Returns the length of the read
     line as an int and a pointer inside the object buffer as char** (so
